Search and replace all instances of unsigned with size_t. -- Test by compiling under Linux 64, Win32, and Win64. -- Win64 found the most errors because their "size_t" is not a "long". -- Required modifying NetCDF-3 C++ interface to resolve issues on Win66.

+++ ocean/test/attenuation_test.cc
@@ -44,7 +44,7 @@ BOOST_AUTO_TEST_CASE( constant_attenuation_test ) {
     // check the answer
 
     double value = 0.01;
-    for (unsigned f = 0; f < freq.size(); ++f) {
+    for (size_t f = 0; f < freq.size(); ++f) {
         BOOST_CHECK_CLOSE(atten(0, 0)(f), value, 1e-6);
         value *= 10.0;
     }
@@ -83,7 +83,7 @@ BOOST_AUTO_TEST_CASE( thorp_test_a ) {
             0.01125, 0.03615, 0.08538, 0.16469, 0.38326, 1.19919, 4.16885,
             12.81169, 27.26378 };
 
-    for (unsigned f = 0; f < freq.size(); ++f) {
+    for (size_t f = 0; f < freq.size(); ++f) {
         cout << freq(f) << "\t" << atten(0, 0)(f) << endl;
         if (freq(f) > 400.0) {
             BOOST_CHECK_CLOSE(atten(0, 0)(f), gsm_thorp[f], 20.0);
@@ -120,7 +120,7 @@ BOOST_AUTO_TEST_CASE( thorp_test_b ) {
     cout << "writing tables to " << name << endl ;
 
     os << "freq,atten" << endl ;
-    for (unsigned f = 0; f < freq.size(); ++f) {
+    for (size_t f = 0; f < freq.size(); ++f) {
         os << freq(f) << "," << atten(0, 0)(f) << endl ;
     }
 }
